distributed/minGPT-ddp/requirements.txt
torch>=1.11.0
hydra-core
